requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pyrebase4==4.7.1
elevenlabs==0.2.24
crewai==0.11.0
gunicorn==21.2.0 